refactor: change Enum type's name. and add comment on OperatorType.
기존에 CalculatorType으로 정의되어 있던 열거 타입의 이름을 OperatorType으로 변경하였다.
연산자와 그 연산을 포함하는 열거 타입일 뿐, 계산기 내지 계산이라는 말과는 관련이 덜하다고 판단하였기 때문이다.
(Calculator 클래스의 존재로 인해 헷갈릴 수 있다는 판단이 들었다)
추가로, 문서화 주석을 각 메서드 및 클래스에 기재하였다. 일부 부족한 부분은 추후 보충할 예정이다.
